Move geometry utilities into their own file
We don't *only* deal with gpkg geometries. And we'll need more of these
as we optimise algorithms
